Moves defaultHandler, defaultResponder to lib
