+++ packages/lib/server/defaultHandler.ts
