+++ packages/lib/server/index.ts
@@ -1,2 +1,3 @@
 export { default as defaultHandler } from "./defaultHandler";
 export { default as defaultResponder } from "./defaultResponder";
+export { getTranslation } from "./i18n";
